Renamed ext dir in Rollbar Dart to extension.
