+++ rollbar_dart/lib/src/extension/trace.dart
