+++ rollbar_dart/test/trace_test.dart
@@ -1,4 +1,4 @@
-import 'package:rollbar_dart/src/ext/trace.dart';
+import 'package:rollbar_dart/src/extension/trace.dart';
 import 'package:test/test.dart';
 
 void main() {
